requirements.txt
# LLM and Agent Framework
openai>=1.52.0
langchain>=0.3.0
langchain-openai>=0.2.0
# langgraph>=0.2.30  # Optional - not used in current ReAct implementation

# Configuration
python-dotenv>=1.0.1

# Chemistry and Materials
pymatgen>=2024.9.16

# Utilities
requests>=2.31.0
pydantic>=2.0.0

# Data Analysis and Evaluation
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# PDF Processing
pypdf>=4.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
